trunk/projetos/ap-comum/source-teste/br/com/ap/comum/conversor/instancia/ConversorDeStringParaStringTest.java
/*
 * ConversorDeStringParaStringTest.java
 * 
 * Data de criação: 03/07/2008
 *
 * 
 * 
 */
package br.com.ap.comum.conversor.instancia;

import org.junit.Test;

import br.com.ap.comum.TesteAbstrato;
import br.com.ap.comum.conversor.instancia.IConversor;
import br.com.ap.comum.excecao.ConversorException;

/**
 * JUnit.
 * 
 * @author AdrianoP
 * @see ConversorDeStringParaString
 */
public class ConversorDeStringParaStringTest extends TesteAbstrato {

	/**
	 * Test method for
	 * {@link br.com.ap.comum.conversor.instancia.ConversorDeStringParaString#converter(java.lang.String)}.
	 * 
	 * @throws ConversorException Exceção de conversão
	 */
	@Test
	public final void testConverter() throws ConversorException {
		IConversor<String, String> c = getConversorFactory()
				.novoConversorDeStringParaString();
		String param = getStringFactory().novaString("35");

		String res = c.converter(param);
		assertEquals(c.getTipoDeOrigem(), String.class);
		assertEquals(c.getTipoDeDestino(), String.class);
		assertNotNull(res);
		assertEquals("35", res);

		res = c.converter(null);
		assertNull(res);
	}

}
